refactor: move view-layer formatting to components
- Move transaction formatting logic from service to component layer
- Relocate type mappings and icons to component/lib/types
- Keep TransactionService focused on data fetching
